pyproject.toml
[project]
name = "metal-coils-warehouse-api"
version = "0.1.0"
description = "API для склада рулонов металла"
authors = [
    {name = "Karina Artsueva <k.artsueva@gmail.com>"}
]
requires-python = ">=3.10"

dependencies = [
    "pydantic-settings",
    "fastapi (>=0.115.11,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "pytz (>=2025.1,<2026.0)",
]

[project.optional-dependencies]
dev = [
    "mypy (>=1.15.0,<2.0.0)",
    "ruff (>=0.11.0,<0.12.0)",
    "autopep8 (>=2.3.2,<3.0.0)",
    "flake8 (>=7.1.2,<8.0.0)",
]

test = [
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.poetry]
package-mode = false